refactor: Move attention.py into attention/ package, rename attention_dsl to attention/cute_dsl
- attention.py → attention/_core.py (re-exported via attention/__init__.py)
- attention_dsl/cute_dsl/ → attention/cute_dsl/ (consistent with mla/cute_dsl/, fused_moe/cute_dsl/)
- Update relative imports in _core.py (. → ..)
- Update import paths in prefill.py and test_cute_dsl_fmha_prefill.py
