+++ flashinfer/attention/_core.py
@@ -20,9 +20,9 @@ from typing import Optional, Tuple, Union
 
 import torch
 
-from .api_logging import flashinfer_api
-from .jit import gen_batch_attention_module
-from .utils import (
+from ..api_logging import flashinfer_api
+from ..jit import gen_batch_attention_module
+from ..utils import (
     MaskMode,
     PosEncodingMode,
     TensorLayout,
@@ -30,9 +30,9 @@ from .utils import (
     _unpack_paged_kv_cache,
     determine_attention_backend,
 )
-from .prefill import BatchPrefillWithPagedKVCacheWrapper
-from .jit.attention.variants import attention_sink_decl
-from .jit.utils import filename_safe_dtype_map
+from ..prefill import BatchPrefillWithPagedKVCacheWrapper
+from ..jit.attention.variants import attention_sink_decl
+from ..jit.utils import filename_safe_dtype_map
 
 
 @functools.cache
